SBN-DataInfra-YouTube-Cookie-Manager-main/requirements.txt
cachetools==5.3.1
certifi==2023.5.7
charset-normalizer==3.1.0
google-api-core==2.11.1
google-auth==2.21.0
google-cloud-secret-manager==2.16.1
googleapis-common-protos==1.59.1
greenlet==2.0.2
grpc-google-iam-v1==0.12.6
grpcio==1.56.0
grpcio-status==1.56.0
idna==3.4
onetimepass==1.0.1
playwright==1.35.0
proto-plus==1.22.3
protobuf==4.23.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
pyee==9.0.4
requests==2.31.0
rsa==4.9
six==1.16.0
typing_extensions==4.6.3
urllib3==1.26.16
